Store LDraw geometry in double precision
Carry full precision end to end:

- read coordinates with -doubleValue instead of -floatValue, and let
  outputStringForFloat: take a double, so neither end of the path narrows;
- widen the MatrixMath geometry types (Point2/3, Vector2/3/4, Size2, Box2/3,
  Matrix3, Matrix4, TransformComponents) to double;
- widen the three directives that keep their own copy of a transform --
  LDrawPart, LDrawLSynth, LDrawRelatedParts. These matter as much as MatrixMath:
  each holds a flat [16] array, so widening the matrix type alone would still
  have narrowed the value back to float on its way out through -write.

Float is kept deliberately at the GPU boundary.

Two hazards here raise no build error, because C narrows double to float
silently. `swift build -Xcc -Wimplicit-float-conversion` surfaces them:

- float32 math functions receiving doubles (roundf, asinf, atan2f, fabsf)
  compute in float and return float, so values round-tripped through float32
  regardless. They sat on the grid-snapping and rotation-decompose paths -- the
  exact places the drift is visible. Swapped for the double versions.

- FloatsApproximatelyEqual was a float32 ULP bit-trick, comparing an int32
  reinterpretation within 5 ULPs. Preserving the mechanism would have made it
  ~1e9 times stricter and quietly broken vertex merging, so it now preserves the
  old tolerance instead: 5 * FLT_EPSILON relative, with the same SMALL_NUMBER
  fallback near zero. That also drops a copy-paste bug in its near-zero branch,
  which tested float1 twice and never float2. Outcomes are unchanged.

Note that LDrawPart's NSCoding payload is now 16 doubles rather than 16 floats.
That is the clipboard and drag-and-drop format, so a copy made by an older build
cannot be pasted into this one.

+++ trunk/Packages/LDrawCore/Sources/LDrawCore/LDrawDrawableElement.m
@@ -322,9 +322,9 @@
 - (Point3) position:(Point3)position
 	  snappedToGrid:(float)gridSpacing
 {
-	position.x = roundf(position.x/gridSpacing) * gridSpacing;
-	position.y = roundf(position.y/gridSpacing) * gridSpacing;
-	position.z = roundf(position.z/gridSpacing) * gridSpacing;
+	position.x = round(position.x/gridSpacing) * gridSpacing;
+	position.y = round(position.y/gridSpacing) * gridSpacing;
+	position.z = round(position.z/gridSpacing) * gridSpacing;
 	
 	return position;
